% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareCrimes.R
\name{catLevels}
\alias{catLevels}
\title{Make levels for merging category predictors}
\usage{
catLevels(levs)
}
\arguments{
\item{levs}{levels of a catagorical variable (factor)}
}
\value{
levels for a new categorical variable of form f1:f2
}
\description{
Make levels for merging category predictors
}
\keyword{internal}

