% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hierclustering.R
\name{clusterPath}
\alias{clusterPath}
\title{Follows path of one crime up a dendrogram}
\usage{
clusterPath(crimeID, tree)
}
\arguments{
\item{crimeID}{the crime ID for a crime used in hierarchical clustering}

\item{tree}{an object produced from \code{\link{crimeClust_hier}}}
}
\value{
data.frame of the additional crimes and the log Bayes factor at each
   merge.
}
\description{
The sequence of groups that a crime belongs to.
}
\details{
Agglomerative hierarchical clustering form clusters by sequentially
   merging the most similar groups at each iteration. This function is designed
   to help trace the sequence of groups an individual crime is a member of. And
   it shows at what score (log Bayes factor) the merging occurred.
}
\examples{
# See vignette: "Crime Series Identification and Clustering" for usage.
}
\seealso{
\code{\link{crimeClust_hier}, \link{plot_hcc}}
}

