% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{color}
\alias{color}
\title{Creates transparent colors}
\usage{
color(col, alpha = NULL)
}
\arguments{
\item{col}{Color that \R recognizes (names or number)}

\item{alpha}{transparency value}
}
\description{
Creates transparent colors
}
\seealso{
\code{\link{col2rgb}}
}
\keyword{internal}

