% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareCrimes.R
\name{compareCategorical}
\alias{compareCategorical}
\title{Make evidence variables from categorical crime data}
\usage{
compareCategorical(C1, C2, levs, binary = FALSE)
}
\arguments{
\item{C1}{length n categorical values of crime attributes}

\item{C2}{length n categorical values of crime attributes}

\item{levs}{the levels of all possible values}

\item{binary}{(logical) match/no match or all combinations}
}
\value{
if binary=TRUE: 1 for match, 0 for non-matches;
 if binary=FALSE: factor vector of merged values (in form of f1:f2)
}
\description{
Compares categorical crime data to check if they match.
}
\keyword{internal}

