% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareCrimes.R
\name{compareNumeric}
\alias{compareNumeric}
\title{Make evidence variables from numeric crime data}
\usage{
compareNumeric(C1, C2)
}
\arguments{
\item{C1}{length n numerical values of crime attributes}

\item{C2}{length n numerical values of crime attributes}
}
\value{
numeric vector of absolute differences
}
\description{
Calculates absolute difference between crimes variables
}
\keyword{internal}

