% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compareCrimes.R
\name{compareTemporal}
\alias{compareTemporal}
\title{Make temporal evidence variable from (possibly uncertain) temporal info}
\usage{
compareTemporal(DT1, DT2, niters = 2000, show.pb = TRUE)
}
\arguments{
\item{DT1}{(n x 2) data.frame of (DT.FROM,DT.TO) for the crimes}

\item{DT2}{(n x 2) data.frame of (DT.FROM,DT.TO) for the crimes}

\item{niters}{number of iterations for Monte Carlo expectation}

\item{show.pb}{(logical) show the progress bar}
}
\value{
data.frame of expected absolute differences:
   \itemize{
     \item temporal - overall difference (in days)  [0,max]
     \item tod - time of day difference (in hours)  [0,12]
     \item dow - fractional day of week difference (in days) [0,3.5]
   }
}
\description{
Calculates the temporal distance between crimes
}
\details{
Uses Monte Carlo expected value - aoristic like analysis
}
\keyword{internal}

