% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crimes}
\alias{crimes}
\title{Ficticious dataset of crime events}
\format{490 crime events
\describe{
   \item{crimeID}{The crime ID number}
   \item{X, Y}{Spatial coordinates}
   \item{MO1}{A categorical MO variable that takes values {1,\dots,31}}
   \item{MO2}{A categorical MO variable that takes values {a,\dots,h}}
   \item{MO3}{A categorical MO variable that takes values {A,\dots,O}}
   \item{DT.FROM}{The earliest possible Date-time of the crime.}
   \item{DT.TO}{The latest possible Date-time of the crime}
 }}
\source{
Ficticious data, but hopefully realistic
}
\usage{
data(crimes)
}
\description{
Some realistic, but fictious, crime incident data.
}
\examples{
head(crimes)
}
\keyword{datasets}

