% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/caseUtils.R
\name{dtdiff}
\alias{dtdiff}
\title{Calculates time between two vectors of datetimes}
\usage{
dtdiff(t1, t2, units = "days")
}
\arguments{
\item{t1}{first set of times}

\item{t2}{second set of times}

\item{units}{one of: "auto", "secs", "mins", "hours","days", "weeks"}
}
\value{
numeric vector of times between the datetime objects
}
\description{
Calculates time between two vectors of datetimes
}
\keyword{internal}

