% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareCrimes.R
\name{expAbsDiff}
\alias{expAbsDiff}
\title{Expected absolute difference of two uniform RVs}
\usage{
expAbsDiff(X, Y)
}
\arguments{
\item{X}{c(min,max)}

\item{Y}{c(min,max)}
}
\value{
the expected absolute difference
}
\description{
Calculates the expected absolute difference of two uniform rv's
}
\keyword{internal}

