% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareCrimes.R
\name{getD}
\alias{getD}
\title{Expected absolute distance of a circular uniform RV to a point}
\usage{
getD(y, X, mod = 24)
}
\arguments{
\item{y}{vector of times in [0,mod)}

\item{X}{c(min,min+length). X[1] must be >= 0 and X[2] >= X[1]. It is possible
 that X[2] can be > mod. I.e., do not do \code{X \%\% mod}}

\item{mod}{the period of time. E.g., mod=24 for time of day (in hours),
 mod=7 for day of week (in days)}
}
\value{
the expected absolute difference
}
\description{
Calculates the expected circular distance between a uniform rv and a point
}
\keyword{internal}

