% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{make.breaks}
\alias{make.breaks}
\title{Make break points for binning continuous predictors}
\usage{
make.breaks(x, type = "quantile", nbins = NULL, binwidth = NULL)
}
\arguments{
\item{x}{observed sample}

\item{type}{one of \code{width} (fixed width) or \code{quantile} binning}

\item{nbins}{number of bins}

\item{binwidth}{bin width; corresponds to quantiles if type='quantile'}
}
\value{
set of unique break points for binning
}
\description{
Make break points for binning continuous predictors
}
\keyword{internal}

