% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{offenders}
\alias{offenders}
\title{Ficticious offender data}
\format{1357 offenders committed 1377 crimes
\describe{
 \item{offenderID}{ID number of offender}
 \item{crimeID}{ID number of crime}
}}
\source{
Ficticious data, but hopefully realistic
}
\usage{
data(offenders)
}
\description{
Offender table relating crimes (\code{crimeID}) to offenders (\code{offenderID})
}
\examples{
head(offenders)
}
\keyword{datasets}

