% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{plotBKG}
\alias{plotBKG}
\title{Generate a background plot}
\usage{
plotBKG(xlim, ylim, background = TRUE, x.minor, x.major, grid.lines = TRUE,
  glwd1 = 2, glwd2 = 1, bkg.col = "grey90", grid.col = "white",
  boxed = TRUE, ...)
}
\arguments{
\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{background}{(logical) should a background color be used}

\item{x.minor}{values for minor axis}

\item{x.major}{values for major axis}

\item{grid.lines}{(logical) should grid lines be added to plot}

\item{glwd1}{linewidth of gridlines}

\item{glwd2}{linewidth of gridlines}

\item{bkg.col}{color of plot background}

\item{grid.col}{color of grid lines}

\item{boxed}{(logical) should plot be boxed}

\item{\ldots}{other arguments passed to plot}
}
\value{
makes a plot
}
\description{
This facilitates a common plot background
}
\keyword{internal}

