% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{predictBF}
\alias{predictBF}
\title{Generate prediction of a component bayes factor}
\usage{
predictBF(BF, x, log = TRUE)
}
\arguments{
\item{BF}{bayes factor data.frame from \code{\link{getBF}}}

\item{x}{vector of new predictor values}

\item{log}{(logical) if \code{TRUE}, return the \bold{log} bayes factor estimate}
}
\value{
estimated (log) bayes factor from a single predictor
}
\description{
This does not include the log prior odds, so will be off by a constant
}
\examples{
# See vignette: "Statistical Methods for Crime Series Linkage" for usage.
}

