% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{predictNB}
\alias{predictNB}
\title{Generate prediction (sum of log bayes factors) from a \code{naiveBayes} object}
\usage{
predictNB(NB, X, components = FALSE, vars = NULL)
}
\arguments{
\item{NB}{a naive bayes object from \code{\link{naiveBayes}}}

\item{X}{data.frame of new predictors, column names must match NB names}

\item{components}{(logical) return the log bayes factors from each component
  or return the sum of log bayes factors}

\item{vars}{the names or column numbers of specific predictors. If NULL, then
  all predictors will be used}
}
\value{
BF if \code{components = FALSE}, the sum of log bayes factors, if
   \code{components = TRUE} the component bayes factors (useful for plotting)
}
\description{
This does not include the log prior odds, so will be off by a
   constant
}
\examples{
# See vignette: "Statistical Methods for Crime Series Linkage" for usage.
}

