% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naivebayes.R
\name{plotBF}
\alias{plotBF}
\title{plots 1D bayes factor}
\usage{
plotBF(BF, log.scale = TRUE, show.legend = TRUE, xlim, ylim = NULL,
  cols = c(color("darkred", alpha = 0.75), color("darkblue", alpha = 0.75)),
  ...)
}
\arguments{
\item{BF}{Bayes Factor}

\item{log.scale}{(logical)}

\item{show.legend}{(logical)}

\item{xlim}{range of x-axis}

\item{ylim}{range of y-axis}

\item{cols}{Colors for plotting. First element is for linkage, second unlinked}

\item{\ldots}{arguemnts passed into \code{\link{plotBKG}}}
}
\value{
plot of Bayes factor
}
\description{
plots 1D bayes factor
}
\examples{
# See vignette: "Statistical Methods for Crime Series Linkage" for usage.
}
\seealso{
\code{\link{plot.naiveBayes}}, \code{\link{plotBKG}}
}

